100: Batched repo select code review fixes
- Disabled linting for a console.warn
- Changed SELECT_REPOSITORY_FORM_TOGGLE_REPOSITORY action value
- Moved font declaration to variables
- Relocated existing button component to vanilla directory
- Some codestyle fixes
