--- src/common/components/vanilla/button/button-positive.css
@@ -1,29 +1,0 @@
-$button-color: $success;
-$button-border-color: $success;
-
-.button-positive {
-  composes: button from './button.css';
-  background-color: $button-color;
-  border-color: $button-border-color;
-  border-radius: .125rem;
-  border-style: solid;
-  border-width: 1px;
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-  color: #fff;
-  cursor: pointer;
-  display: inline-block;
-  font-family: "Ubuntu", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-  font-size: 1rem;
-  font-smoothing: subpixel-antialiased;
-  font-weight: 300;
-  line-height: 1rem;
-  margin: 0;
-  outline: none;
-  padding: .6875rem 1.5rem;
-  text-align: center;
-  text-decoration: none;
-  -webkit-transition: background-color .2s;
-  transition: background-color .2s;
-  width: auto;
-}
