+++ src/common/components/vanilla/button/button.css
@@ -1,3 +1,89 @@
+$primary-button-color: $ubuntu-orange;
+$primary-button-color-hover: #d44615;
+$primary-button-text-color: #FFF;
+
+$secondary-button-color: #FFF;
+$secondary-button-color-hover: #EFEFEF;
+$secondary-button-text-color: $ubuntu-orange;
+$secondary-border-color: $mid-grey;
+
+$positive-button-color: $success;
+$positive-button-color-hover: #0a5615;
+$positive-button-text-color: #FFF;
+$positive-border-color: $success;
+
 .button {
-  font-family: "Ubuntu", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+  font-family: $base-font-family;
+  font-smoothing: subpixel-antialiased;
+  font-size: 1em;
+  display: inline-block;
+  margin: 0;
+  padding: 6px 19px;
+  text-align: center;
+  text-decoration: none;
+  border: 0;
+  border-radius: 2px;
+  outline: none;
+  font-weight: 300;
+  cursor: pointer;
+}
+
+.button:active,
+.button:focus,
+.button:hover {
+  text-decoration: none;
+}
+
+.primary {
+  composes: button;
+
+  color: $primary-button-text-color;
+  background-color: $primary-button-color;
+}
+
+.primary:active,
+.primary:focus,
+.primary:hover {
+  background-color: $primary-button-color-hover;
+}
+
+.secondary {
+  composes: button;
+
+  color: $secondary-button-text-color;
+  background-color: $secondary-button-color;
+  border: 1px solid $secondary-border-color;
+}
+
+.secondary:active,
+.secondary:focus,
+.secondary:hover {
+  background-color: $secondary-button-color-hover;
+}
+
+.positive {
+  composes: button;
+
+  color: $positive-button-text-color;
+  background-color: $positive-button-color;
+  border: 1px solid $positive-border-color;
+}
+
+.positive:active,
+.positive:focus,
+.positive:hover {
+  background-color: $positive-button-color-hover;
+}
+
+.button:disabled {
+  cursor: not-allowed;
+  opacity: .5;
+}
+
+.primary:disabled:hover {
+  background-color: $primary-button-color;
+}
+
+.secondary:disabled:hover {
+  background-color: $secondary-button-color;
 }
