+++ src/common/components/vanilla/heading/heading.css
@@ -1,5 +1,5 @@
 .heading {
-  font-family: "Ubuntu", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+  font-family: $base-font-family;
   font-weight: 300;
   margin-bottom: 1.2rem;
   margin-top: 0;
