+++ src/common/helpers/repositories.js
@@ -1,5 +1,5 @@
 export function hasRepository(arr, repository) {
   return arr.filter((item) => {
-    return item.fullName === repository.fullName;
+    return (item.fullName === repository.fullName);
   }).length > 0;
 }
