+++ src/common/routes.js
@@ -13,10 +13,10 @@ export default (
   <Route component={App}>
     <Route path="/" component={Landing}/>
     <Route path="/dashboard" component={Dashboard}/>
+    <Route path="/dashboard/select-repositories" component={SelectRepositories}/>
     <Route path="/:owner/:name/setup" component={RepositorySetup}/>
     <Route path="/:owner/:name/builds" component={Builds}/>
     <Route path="/:owner/:name/builds/:buildId" component={BuildDetails}/>
     <Route path="/login/failed" component={LoginFailed}/>
-    <Route path="/dashboard/select-repositories" component={SelectRepositories}/>
   </Route>
 );
